Fix calculator operator button function logic
problem: first, the value1 and value2 is stored
when the operator button is clicked. it flaws.

solution(roughly):
when user is click the button,
store the current display text
in value1 variable immediately.
then if the operator button is clicked,
make the boolean value isOperatorButtonClicked
to true, then save the operator button value,
then when user click the number again,
because the operatorbutton condition is true,
the current text display is saved to value2
variable, then reassign the value1 variable value
with function operate().